requirements.txt
Pillow==10.3.0
transformers==4.41.2
